third_party/blink/web_tests/fast/xsl/xslt-multiple-relative-import-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
crbug.com/436479: Test for multiple xsl:import with relative URLs.

SUCCESS
